% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAffyExpr.R
\name{getAffyExpr}
\alias{getAffyExpr}
\title{Create TARGET OS expression SummarizedExperiment from raw data}
\usage{
getAffyExpr()
}
\description{
Create TARGET OS expression SummarizedExperiment from raw data in the TARGETOsteoDataPackage
}

