% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinical.R
\name{getClinical}
\alias{getClinical}
\title{Create TARGET OS clinical data.frame from raw data}
\usage{
getClinical()
}
\description{
Create TARGET OS clinical data.frame from raw data in the TARGETOsteoDataPackage
}

