% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCNSeg.R
\name{getCNSeg}
\alias{getCNSeg}
\title{Create TARGET OS clinical data.frame from raw data}
\usage{
getCNSeg()
}
\description{
Create TARGET OS clinical data.frame from raw data in the TARGETOsteoDataPackage
}

