% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethylation.R
\name{getMethylation}
\alias{getMethylation}
\title{Create TARGET OS clinical data.frame from raw data}
\usage{
getMethylation()
}
\description{
Create TARGET OS clinical data.frame from raw data in the TARGETOsteoDataPackage
}

