% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmiRNA.R
\name{getmiRNA}
\alias{getmiRNA}
\title{Create TARGET OS clinical data.frame from raw data}
\usage{
getmiRNA()
}
\description{
Create TARGET OS clinical data.frame from raw data in the TARGETOsteoDataPackage
}

